Oleksandr Prognimak; http://jira:8070/browse/TPI-48;https://dev.crowdcode.io/jira/browse/ISDTPI-52; Refactor to use without AOP
