+++ applauncher/src/main/java/de/interseroh/tmb/applauncher/server/ApplauncherApplication.java
@@ -21,6 +21,7 @@ package de.interseroh.tmb.applauncher.server;
 import com.google.gwt.logging.server.RemoteLoggingServiceImpl;
 import de.interseroh.tmb.applauncher.shared.ApplauncherServiceEndpoint;
 import de.interseroh.tmb.common.CrossOriginLoggingFilter;
+import de.interseroh.tmb.common.LoggingCrossOriginConfiguration;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
@@ -28,22 +29,23 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Import;
 
+import javax.servlet.FilterRegistration;
 import javax.servlet.ServletContext;
 
-@Aspect
+
 @SpringBootApplication
-@ComponentScan(basePackages = {"de.interseroh.tmb.common"})
+@Import(LoggingCrossOriginConfiguration.class)
 public class ApplauncherApplication {
 
     @Value("${server.context-path}")
     private String contextPath;
 
-    @Autowired
-    private CrossOriginLoggingFilter loggingFilter;
 
     public static void main(String[] args) {
         SpringApplication.run(ApplauncherApplication.class, args);
@@ -55,9 +57,6 @@ public class ApplauncherApplication {
                 contextPath.concat(ApplauncherServiceEndpoint.GWT_REMOTE_LOGGING) + "/*");
     }
 
-    @Before("execution( * org.springframework.boot.web.servlet.ServletRegistrationBean.onStartup(..))  ")
-    public void addFilter(JoinPoint joinPoint) {
-        ServletContext ctx = (ServletContext) joinPoint.getArgs()[0];
-        ctx.addFilter("loggingFilter", loggingFilter);
-    }
+
+
 }
