+++ topmenubar/src/main/java/de/interseroh/tmb/server/TopMenuBarApplication.java
@@ -22,7 +22,7 @@ import de.interseroh.tmb.shared.TopMenueBarServiceEndpoint;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.embedded.ServletRegistrationBean;
+import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.annotation.Bean;
 
 import com.google.gwt.logging.server.RemoteLoggingServiceImpl;
